tast-tests/webrtc: speed up MediaRecorder test case
MediaRecorder is still flaky on slows devices. A deeper investigation
revealed that the issue is that MediaRecorders that have been started
in JS are not immediately garbage-collected, and keep encoding whatever
video/audio is being fed to them (the fake video capture device and
its associated 'beep'). This made the test progressively slower and
led to timeouts in slow devices.

This test fixes that by stop()ping any |recorder| that has been started
and making the Fake video capture device one half -- since this is an
API test, the size of the encoding is irrelevant and smaller means
faster.

With this changes and on my nocturne celeron (slowest I have at hand)
the JS run goes from taking about 20s to taking about 8s.

BUG=b:158849502
TEST=tast run DUT webrtc.MediaRecorder on nocturne and celes

Change-Id: Ie71b994b2fb7461a2ef2ed43886a0bca83617cb6