wifi: SimpleConnect: Change the channel of 80211acvht40 case.
The veyron_mickey DUTs in lab have some problems with channel 120. As
this test is for testing the VHT 40MHz setting, select a more commonly
used channel for it and move the channel 120 test to a new DFS testcase
so that the ported tests can start replacing the Autotest counterparts.

BUG=b:154440798
TEST=tast run $DUT wifi.SimpleConnect.80211acvht40
TEST=tast run $DUT wifi.SimpleConnect.dfs_ch120

Change-Id: Idd6c678462066d93af957124f4a0e7108686ba64