Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Hmm, I'm a bit surprised that we're keeping the extension key secret. I have a few questions:
> 
> - What was the answer to https://bugs.chromium.org/p/chromium/issues/detail?id=941107#c6 ? It looks unanswered.
oops, just noticed that (I inherited the task from the previous owner). We ended up with a new command line flag specifically for signin profile.

> - Can we avoid relying on secrecy of an extension private key, just like existing Autotest extensions? If not, what is the difference?
main diff is we load that extension in the signin profile. and want to make it as hard as possible to use it as attack surface (e.g. we only allow it for test images). The id whitelisted specifically for signin profile + test images. (https://source.chromium.org/chromium/chromium/src/+/master:extensions/common/constants.cc;l=147)

This solution came from discussion in https://crrev.com/c/1521272
"I'm a little nervous that we're allowlisting capabilities for an ID that is used in all kinds of browser tests, since at least the key to generate an extension with that ID is public here in the source, and I wouldn't be surprised if the PEM is included as well.  Even though we don't make strong security guarantees about unpacked extensions or out-of-store extensions, I'd feel better if the allowlisted extension was one that we controlled the PEM for outside of the Chromium repo, and wasn't conflated with an extension used frequently in unrelated test data."

> - What happens if the secret key is accidentally leaked?
Not much we could do I guess. Could always create a new one. But released Chrome would still be affected.

Patch-set: 4
Label: Commit-Queue=0