wifi: CSA: Turn off powersave during the test.
Currently, the action frames sent by FrameSender are not buffered for
DTIM. If the DUT is in powersave mode, it cannot properly receive the
CSA frame and will fail the test. In Autotest, powersave mode is turned
off in warmup so it is not affected. Turn off powersave in wifi.CSA
test to sync the behavior of Autotest version for now so that we can
start replacing the Autotest counterpart.

BUG=b:154879577
TEST=tast run $DUT wifi.CSA on krane with powersave on

Change-Id: I1020497d79756d0df00245b0215cea78607e041a