Update patch set 2
Patch Set 2:

ImportPEMPrivateKeyAndCertBySlot is not the only portion of code that hwsec is exporting to wifi, there's also MountVault(), Unmount(), RemoveVault(), WaitForUserToken() and GetTokenInfoForUser() from common/hwsec. From my offline conversation with Leo, if I understand correctly, is that these APIs from common/hwsec is slated for refactor with the move to use UserDataAuth API after UserDataAuth and Distributed Mode is both fully launched, and that is probably sometime in Q4, but from the impression I get from discussion with Dean Liao, it seemed that this is a bit more urgent than that.

Additionally, such a refactor/move to gRPC doesn't change the paradigm of the API in a significant way because it is mostly internal to our library, and neither does it change set of way WiFi would use TPM APIs, so L80-L110 of crrev/c/2172566/14 will look pretty similar to the way it is now, the biggest difference is probably L162-L168 will be gone.

Patch-set: 2