Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> ImportPEMPrivateKeyAndCertBySlot is not the only portion of code that hwsec is exporting to wifi, there's also MountVault(), Unmount(), RemoveVault(), WaitForUserToken() and GetTokenInfoForUser() from common/hwsec. From my offline conversation with Leo, if I understand correctly, is that these APIs from common/hwsec is slated for refactor with the move to use UserDataAuth API after UserDataAuth and Distributed Mode is both fully launched, and that is probably sometime in Q4, but from the impression I get from discussion with Dean Liao, it seemed that this is a bit more urgent than that.
>
Sorry maybe I didn't clarify enough during the offline discussion, and sorry for being straight this time: what you mentioned was one of those off-topic and userdataauth is NEVER a blocker in this context, i.e., switching to gRPC (and how is this related to distributed model? you are not using any of API changed by distributed model), To me this whole paragraph seems just irrelevant to this CL stack.

And I think you actually have more flexibility than you think, John. If you really think something that is planned happen in Q4 is blocking the best practice you can provide to wifi team (though I can't think of any), you can always promote the priority, start working on it, and finish it ASAP. 
 
> Additionally, such a refactor/move to gRPC doesn't change the paradigm of the API in a significant way because it is mostly internal to our library, and neither does it change set of way WiFi would use TPM APIs, so L80-L110 of crrev/c/2172566/14 will look pretty similar to the way it is now, the biggest difference is probably L162-L168 will be gone.

Sure, it is possible that gRPC is not the best solution here. So, a very straight question to you, John -- if you are pushing back using gRPC, then what is your proposition to address Brian's concern (they are also mine)? Or the current CL is what you think the best? (Reminder: Please do check all comments in this whole CL stack before answering.)

Patch-set: 2