tast-tests: Add hwdep.[No]ForceDischarge to CCAUIPreviewPowerPerf
On some platforms, force discharging a battery with ectool is not
supported. Depending on whether this functionality is supported
or not, it should be decided to use metrics from battery.
This CL parameterizes CCAUIPreviewPowerPerf to decide whether
battery metrics should be collected on a device or not.
It also prevents calling ectool on devices that do not support
force discharge, which leads the test to fail.

BUG=b:157512297
TEST=tast run <DUT w/o battery> camera.CCAUIPreviewPowerPerf* did not
fail

Change-Id: I2000f7baebbe40ad1dfabb5c41ee643167c54712
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2256526
Reviewed-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Matthias Springer <springerm@chromium.org>
Tested-by: Hikaru Nishida <hikalium@chromium.org>
Commit-Queue: Hikaru Nishida <hikalium@chromium.org>