Update patch set 1
Patch Set 1:

By seeing https://tastboard.corp.google.com/test/platform.DLCService, it fails around 13278-13280, due to "DLC combination tests: Install failed: failed to install: exit status 70, Mimic device reboot tests: Install failed: failed to install: exit status 70".
I believe this is known and you already fixed the issue -- that's why it's green now -- but I just want to confirm this is like that.

If this is a fixed one or it turned out unrelated to this test scenario, this is good to go.

Patch-set: 1