Update patch set 6
Patch Set 6:

(2 comments)

I am fine with this change if needed, but in general, I want to avoid tests that install apps via play store. It is a strong source of flakiness and a way to ensure that a test can't be promoted to CQ.

With that in mind, I would generally suggest against trying to add new methods for installation. I made a comment on your other CL about what I would do instead to avoid the play store install.

Patch-set: 6