Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > > Patch Set 2: Commit-Queue+2
> > > > 
> > > > Chrome CL is landing soon. I don't think there's a way to sync them up, the test will fail one way or another for a while. cc this weeks gardener FYI
> > > 
> > > Is not normally landing chrome code first?
> > 
> > If I land chrome code first, this test will break because the metrics will be stale. If I land this first, this test will break because the metrics will be too new. So either way this test will break.
> 
> Feel free to stop the commit if you feel otherwise, I dont have a problem either way.

Please be sure that after a CL is landed to chromium, you'll have to wait for (at least) a few days, until the change is included by the next chrome uprev.

If you really care about the transition, the ideal solution would be to create a change on tast which supports both version of chrome, and then remove the support of the old code after the chrome uprev contains the change.

That said, I think this is also fine, as you expect test failures. Actually this does not break anything -- the failures of performance tests (group:crosbolt) will be mostly ignored and they won't close any trees.  Only thing we can notice would be the reports of the testmon alerts.

Patch-set: 2