Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (2 comments)
> 
> So, I left some comments on the doc. I am ok with adding more tests to the CQ, even if they have funny names and if we just cover a few well chosen models. My main issue is on the workflow. Most regressions come in via Chrome. But tests will hold Chrome uprev and bisector is unusable until Chrome uprevs. We also just have eve an kevin running hwtests in the Chromium waterfall. Do you expect to monitor those? Do you expect to limit the whitelist to these two boards? How are Chrome regressions handled in this world?
> 
> I am going to be OOO until Tuesday next week.

Thanks for all the feedback Ilja! It would be great if we could somehow get these tests in the Chromium waterfall, but I'm not at all familiar with the hwtests. I'm a bit limited on time right now, so if anyone could take this up that would be great. For now let's focus on CQ, even if the Chrome uprev process is not ideal, and see how that goes.

Patch-set: 6