Reduce ui.UbertrayOpenSettings flakiness by using poll-based clicking
When opening the Ubertray, we're often clicking the status area too
quickly, before it's able to receive the click and open the Ubertray.
The Settings button in the Ubertray also has a similar problem. This can
be solved by repeating the click until a certain condition is met, such
as the Ubertray opening or closing. I added a function to automation.go
for this purpose and updated the clicks in ui.UbertrayOpenSettings,
which seem to be quite timing-sensitive.

BUG=chromium:1094086
TEST=tast run localhost:2222 ui.UbertrayOpenSettings

Change-Id: Ifbb6d2f7b9c1b2aa227ab5e770a53f0ec0b7c2b1