camera: Show chart on tablet before running HAL3Remote test.
The HAL3 remote test which is designed to be run in camerabox setup.
Each camerabox setup have a pair of chart tablet and a DUT whose target
facing(front or back) camera to be tested will face to screen of chart
tablet. The test camera's FOV on DUT will cover chart screen which
provide scene controllable environment for DUT's camera.

This CL add a new chart service which will be run on chart tablet to
provide API for displaying arbitrary chart on it. And display a default
chart on it before running the HAL3Remote test.

BUG=b:152724161
TEST=tast run <DUT> "camera.HAL3Remote.*_back"
TEST=tast run <DUT> "camera.HAL3Remote.*_front"

Change-Id: I9b26688d90776f91ab440d381e1ed01c6f00955c