crostini: save ~500MB disk space in Crostini tests
Crostini tests are unable to run on some boards due to a lack of disk
space.  This change deletes some files that are not needed after use
freeing up 500 MB on download tests and 400MB on artifact tests (both
download and artifact tests now use the same amount of disk space).
These deletions do not affect the ability to run other tests, or to run
the same tests again as the files deleted are both created and consumed
entirely in the precondition.

BUG=chromium:1039403
TEST=Ran crostini test on eve with extra disk space logging

Change-Id: I31ae382b746943a69b6cff5ee46b46448649620d