video: More files for DrawOnCanvas and Contents.
This CL creates two more test files for video.DrawOnCanvas and
video.Contents. The files were generated a script, so for the sake of
documentation and to make the generation of files in the future more
systematic, the script is provided in the README.md.

We also take the chance to:

1) Make sure the videos are 30 seconds (the concern that the videos
   could be too short came up in a previous CL review).

2) Change the colors used for the quadrants to be less "random". Still,
   they can't be too nice so that they're unlikely to ever be the same
   color as an unintended artifact.

3) Change how video-on-canvas.html checks the canvas' edges to be
   consistent with play.go: we go 2 pixels into the image instead of
   checking right at the edge.

BUG=b:159158467
TEST=video.{DrawOnCanvas,Contents} on eve/kukui.

Cq-Depend: chromium:2251385
Change-Id: Ia7f88a15a28c0d7c37303948bebac7ae22bae4a9