video: More files for DrawOnCanvas and Contents.
This CL creates two more test files for video.DrawOnCanvas and
video.Contents. The files were generated a script, so for the sake of
documentation and to make the generation of files in the future more
systematic, the script is provided in the README.md.

We also take the chance to:

1) Make sure the videos are 30 seconds (the concern that the videos
   could be too short came up in a previous CL review).

2) Change the colors used for the quadrants to be less "random". Still,
   they can't be too nice so that they're unlikely to ever be the same
   color as an unintended artifact.

3) Make the non-visible area of the videos contain a color that's
   unexpected by the test so that checking for colors at the edges works
   as intended.

More work is needed to make the color checks at the edges not fail due
to blending artifacts, but this is left for a follow-up (I'm actively
working on this test to stabilize it).

BUG=b:159158467
TEST=video.Contents.h264_1080p_hw on nami fails as expected.

Change-Id: Ia7f88a15a28c0d7c37303948bebac7ae22bae4a9