tast.graphics.sanity: port windowmanagertest to webgl
Drop glbench/windowmanagertest, porting it to webgl displayed in a
fullscreen browser window.

This gets rid of a dependency on waffle, which is completely broken
on qcom platforms, as it does not pass UBWC modifier to kernel.  Rather
than spend time updating our null waffle fork, we should just drop the
use of waffle in graphics.Sanity.  (Waffle is not used anywhere else in
production, and using the real UI for this test makes it operate in
a way that is more similar to real life, which seems like a useful
benefit.)

BUG=b:158238296
TEST=run graphics.Sanity on trogdor

Change-Id: Ib7bf2bd5e139bc9df08f65951be893191d304322