Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> tast-owners@: I'd like to hear from Tast owners about having foo.Bar and foo.Bar.unstable. What we want to do here is to promote stable subtests to CQ earlier while leaving unstable ones in informational. This pattern is emerging in several places recently and we'd like to know whether there is a recommended solution, or whether this solution sounds good to you.

I think this is the common pattern.
The detailed condition what is stable/unstable depends on case-by-case.
We understand the demand to promote some tests to CQ, while keeping others which is something similar in informational to improve the test coverage incrementally.
I hope you'll work on stabilizing the unstable case.

Patch-set: 1