dptf_override_cros_config: skip atlas
Atlas is receiving unibuild configuration, per go/unibuild-migration.

Since Atlas differentiates dptf profile based on memory part instead
of identifiable information in cros_config (which is an obscure design
we don't want to repeat), we need to keep the legacy dptf_override.sh
script in the overlay.

If we don't skip this test for atlas, after the migration, we get this
error message:
    Unexpected DPTF profile: got "/etc/dptf/1098_0987.bin"; want ""

So just skip the test for atlas.

BUG=chromium:1098942
TEST=power.DptfOverrideCrosConfig skips on atlas

Change-Id: I69b7839085a253e999ac0a619546f7b74227c93b