tast-tests: Workaround for activity.IsRunning
activity.IsRunning is broken on Android R, which causes many power tests
to fail. These tests check if an Android app started during the test is
still running by the end of the test. This CL deactivates that check, so
that we can get power numbers until the issue is fixed.

BUG=b:157695805
TEST=tast run DUT arc.PowerCameraPreviewPerf.30fps_vm

Change-Id: Ic26c05c86cbb30c165d6d052798a6895913157b7