platform: storage: change fail case for SATA and NVMe
Since tests show that devices can work beyond the manufacturer
recommendation, we should be more stringent and not look at the
Percentage Used. For SATA we should use attribute 160
Uncorrectable_Error_Cnt or 187 Reported_Uncorrect. For NVMe we
should look at Available Spare and fail if it is below Available
Spare Threshold.

BUG=b:154925099
TEST=Ran platform.StorageWearoutDetect

Change-Id: Iee545b15f1d22bdc36ae069ffbd6232c6e3bc034