crash: Fix race when setting up filter-in.
If there's a crash_reporter process running already when new contents
are written to filter-in, that process won't pick up the changes and
will write its crash out even if it does not match the filter. This
causes issues with tests that rely on no unexpected crashes being
written.

To fix this, have the enableCrashFiltering method wait for
crash_reporter to finish.

BUG=chromium:1098042,b:159816248
TEST=tast run $DUT crash.Sender*
TEST=fast_build.sh -T

Fixed: chromium:1098042,b:159816248
Change-Id: Icc0d620bbbc50232ccd3969cc118de0244781e85