Fix the test failures of skia_renderer precondition
It seems LoggedInWith100DummyApps() precondition and its skia
renderer variation share the same "name", and it seems the same
naming causes a weird failures on preparation.

This CL simply adds "_skia_renderer" at the end of the name when
it is enabled, so that the name won't conflict anymore.

BUG=chromium:1100425
TEST=tast ui 'ui.Launcher*Perf*' 'ui.OverviewPerf*'

Change-Id: I1f3bb9d113e31258c6c2eccb709459593a2c30c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2274060
Tested-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>