Fix crostini precondition on some devices
Some devices in the lab actually do not have physical keyboards,
or the device configuration is in tablet-mode style (i.e. flipped).
In that case, a virtual keyboard may appear and cover the "install"
button of crostini installer, which blocks further operations.

This CL invokes HideVirtualKeyboard so that it won't block the
operations anymore.

This also increases the timeout of crostini preconditions. Because
the installer may have to wait for upto 10 minutes, the timeout
might be too short. When it gets stuck for an unexpected reason,
it runs out the entire timeout of the precondition and further
faillog (like dumping UI-tree) also fails. Increasing all of
preconditions by 4 minutes so that every precondition has at least
10 minutes.

BUG=chromium:1100710
TEST=tast run crostini.AudioSanity'*' on nocturne, without keyboard

Change-Id: Ia76cc4e074999f5305cc3e0b39e67560df8db7c9