video: Fix PlaybackPerf deps for 60fps & 4K.
This CL fixes video.PlaybackPerf to take into account that we have
separate capabilities per codec for 4K and for 60fps videos. For
example, kukui can't HW decode 4K videos and since prior to this CL we
were gating on HWDecodeH264 only, video.PlaybackPerf.h264_2160p_*_hw was
failing.

Thus, for 60fps tests, this CL uses the _60 variations of the
capabilities. For 4K tests, we use the _4K variations.

BUG=b:159802807
TEST=PlaybackPerf.h264_2160p_*_hw is skipped on kukui.

Change-Id: I6240e5fe9ecfdeb8278e104db964fca2894e0de7