Set the timeout for ash.WaitForLauncherState
ash.WaitForLauncherState does not have a timeout, and therefore
it keeps waiting until the context gets expired. In case that
the launcher won't be in the desired state, usually it waits
until the test itself times out. It is better to set a timeout
so that the test can fail more timely.

This CL sets 10 seconds for the timeout; the current usage of
this function is to wait right after some user operations
(shortcut key or mouse), so 10 seconds should be long enough
for launcher state.

BUG=none
TEST=tast run

Change-Id: I9798fbc1fb661f06eaaa4c13cee318d00ccf41a2