Update patch set 2
Patch Set 2:

> I only do the Send method partly due to crbug.com/1030635.

Ah, right. We really need to get an OS update out to our APs...

> And in Tast, ssh execution is much faster than Autotest, so calling Send multiple times might be acceptable for now.

For the existing test (wifi.CSA), this makes total sense. But for some others, we want a continuous stream of incrementing frames. I suppose one *can* do that with a ton of individual SSH calls (or maybe a goroutine that loops over a batch of them?), but that doesn't seem ideal. Anyway, we will cross that bridge when we get there.

Patch-set: 2