tast-tests: hwsec: Change hwsec.MountCombination to run on tpm1.2 devices as well
To port hwsec.MountCombination to run on TPMv1.2 devices, we'll need to
remove the reset system state code (which is not necessary for this test).
Also, to allow it to run on TPMv1.2 devices, we need to remove the low
entropy credential testing because tpm1.2 doesn't support it.
Testing for low entropy credential is covered by
platform_CryptohomeLECredentialManager autotest.

BUG=b:160260730
TEST=tast run $dut hwsec.MountCombination on relm and nami

Change-Id: I62b9ffb8b58dc6e465315e92cd5ce9516ab44b8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2277948
Commit-Queue: John L Chen <zuan@chromium.org>
Tested-by: John L Chen <zuan@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>