Make screenshot before closing the Assistant UI
When a test fails the framework takes a screenshot, to help you debug.
However our tests end by stopping the Assistant (to ensure the next test
is not affected), which closes the UI and makes the screenshot useless.

This change ensures we take the screenshot before the Assistant UI is
closed.

BUG=N/A
TEST=Ran Assistant TAST tests, and ensured the screenshot showed up when the test fails.

Change-Id: I61e9956582767deb46ed08ba42d809a30bfdb2ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2278117
Commit-Queue: Jeroen Dhollander <jeroendh@chromium.org>
Tested-by: Jeroen Dhollander <jeroendh@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Auto-Submit: Jeroen Dhollander <jeroendh@chromium.org>