tast-tests: Add ReadBatteryCapacity/Status and LowBatteryShutdownPercent
Before this CL, checking the battery capacity and status were depending
on old battery metrics mechanism in battery_metrics.go. To remove
battery_metrics.go, we need to reimplement the functions needed for that
checking in sysfs_battery_metrics.go. This CL adds
ReadBatteryCapacity/Status and utilize them for the checking, and move
the LowBatteryShutdownPercent into sysfs_battery_metrics.go to be
prepared for deletion of battery_metrics.go.

BUG=b:159586810
TEST=Kept the behavior before this CL

Change-Id: Ib0960f732059465f92d4560170db46d0ee529882