Update patch set 7
Patch Set 7:

Hmm. "Failed to wait wiping finished: wipe_init failed" error is still concerning if we haven't done anything about it.

Wild guess: is it possible that it's too early after reboot to call gooftool? wipe_init.log says:

 [2020-06-25 00:52:02,464] INFO:root:Running services (service_name, LOG_PATH): [('boot-services', ''), ('failsafe', ''), ('dbus', ''), ('wpasupplicant', ''), ('openssh-server', ''), ('shill', ''), ('sslh', '')]
 [2020-06-25 00:52:02,465] INFO:root:Going to stop services (service_name, LOG_PATH): []

The "going to stop" list is empty, and I guess it's not what you expect (e.g. "ui" will be there usually). It seems to suggest that the system is still booting.

Patch-set: 7