tast-tests/ force use hw decoder in appropriate media preconditions
https://crrev.com/c/2247331 landed a new videostack algorithm to
select hw or sw video decoders; for ChromeOS this meant using sw
decoding for resolutions < 720p. Of course this broke plenty of
video.Play* tests :-) to name a few.

This CL disables the ResolutionBasedDecoderPriority for the appropriate
media preconditions, so that the tests can see their greenness
restored. (We'll tackle in parallel what's the best threshold for
hw or sw video decoder in production).

BUG=b:160343274
TEST=tast -verbose run ${IP} video.Play.*vp8*  on kohaku + the
   https://crrev.com/c/2278702 CL which contains the new flag.

Change-Id: I05f3ccaea856a6124b5e7645d1d11a9f82e74cf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2285027
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>