crostini: Add a performance test for mouse movement
This test evaluates consistency of mouse movements. It moves the mouse
at a steady pace and records the output of an applet which tracks the
mouse. We report the range and standard deviation of movement *deltas*,
which sohuld both be 0 if mouse movements were translated perfectly to
crostini apps (because the test code sends events consistently).

In other words, this test evaluates how imprecise the translation is.

BUG=b:160430880
TEST=tast -verbose run $DUT crostini.MousePerf

Change-Id: I2921f0e1a34a738b9e89c6ef190fb1a22a07f5b0