Reset the automation API feature on Chrome.ResetState()
When the same Chrome instance is reused (i.e. test cases share
the same precondition), it may implicitly share the automation
feature activation. When a test activates it, it stays activated
as long as the test extension terminates.

Even when the feature is not used explicitly, the extension API
emits regarding accessibility events (node updates, location
changes) and the extension keeps the tree nodes in sync, and
that may affect the results, especially when it is a performance
test which involves a lot of accessibility nodes.

This CL works with crrev.com/c/2293183 to ensure that those
event handling will be cleared at ResetState().

BUG=chromium:1096719
TEST=ui.DragTabInClamshellPerf ui.OverviewScrollPerf ui.SystemTrayPerf

Cq-Depend: chromium:2293183
Change-Id: I336d3b4e7612745f24f238c0dfc80a000e848d26
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2290421
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>