tast-tests: syzkaller: enable debugging flag
Going through internal stainless log results, it looks like the syz-manager
is having trouble ssh'ing into the DUT when run as part of the testsuite.
This does not occur when running the tast test against a leased device.

Enable the debugging flag for now in order to diagnoze potential issues.

BUG=b:156755711
TEST=tast -verbose run localhost:2200 syzkaller.Wrapper

Change-Id: I28101cf6dcbae18f684fcd7284830431d94b6fd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2290791
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Zubin Mithra <zsm@chromium.org>
Commit-Queue: Zubin Mithra <zsm@chromium.org>