network/LockCheckNetworkHook: touch lock before unlocking
See $BUG for more info. When a test releases this lock, it doesn't
guarantee that upstream connectivity is completely restored; it only
guarantees that the test has finished with any of its destructive
actions (e.g., it has finished restarting Shill; it has finished
resetting its IP address; it has finished its suspend/resume cycle). The
DUT could still be renegotiating DHCP, for instance.

It's nice if check_ethernet.hook knows not only when the last test actor
has *started* grabbing the lock, but when the lock was last released.
That way, it can provide some grace period for the DUT to finish
restoring its own connectivity, undisturbed.

BUG=chromium:1083044
TEST=run a network test that stops shill, in a loop, while
     simultaneously running check_ethernet.hook in a tight loop;
     we expect not to see check_ethernet.hook triggering any recovery
     actions (and so, never see it log "successful after XX seconds")
     `while : ; do tast -verbose run -failfortests \
        ${HOST} network.DefaultProfile || break; done`
     on DUT:
     `stop recover_duts
      while : ; do
        /usr/local/libexec/recover-duts/hooks/check_ethernet.hook 2>&1 | \
        tee /dev/tty | grep -q 'successful after [0-9]* seconds' && break
      done`

Change-Id: Ic978f781a2c4f8164b756795c24efadbd0be4f45