Sleep between Alt and Tab in ui.WindowCycle
This test is currently quite flaky due to the window cycle menu
occasionally not appearing. ui.WindowCyclePerf performs similar actions
but is far less flaky. One major difference is the presence of sleeps in
between key presses. We can try adding a sleep in between the Alt and
Tab key presses to bring up the cycle menu to see if it improves test
stability. Used the same sleep amount as ui.WindowCyclePerf as a
starting point, but we can try lowering it further after we see if it
fixes the flakiness at all.

BUG=chromium:1081688
TEST=tast run chromeos4-row3-rack8-host10.cros ui.WindowCycle

Change-Id: Iefb64641c55fdf1444c9c3e2385aeed22142347e