Modify NT01 for a better coverage
After revisiting NT01,it was noticed that some cases are not covered.
There are two sets of test-cases, first is for activities that their
orientation is specified, and second for unspecified activities.
For specified activities, there are two test-cases one for landscape
and one for portrait. For each activity the display is set to the
opposite orientation and then the activity will launch. This should
rotate the display to activity’s desired orientation.
For unspecified activities, the display will be set to a given
orientation, then unspecified activity will launch, this shouldn’t
change the display orientation, also the activity orientation will be
asserted to make sure it matches the display orientation.
BUG=b:140203139
TEST=tast run $DUT arc.WMNonresizableCTablet

Change-Id: Ia2be9a1cf36c923d9e30a985c06fdc7519154f6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2292843
Tested-by: Armen Keshishian <armenk@google.com>
Auto-Submit: Armen Keshishian <armenk@google.com>
Commit-Queue: Armen Keshishian <armenk@google.com>
Reviewed-by: Jun Mukai <mukai@chromium.org>