crostini: ui: click 'custom' before adjusting slider in setDiskSize
The slider doesn't exist by default. This CL clicks the 'custom'
radiobox before sending right key to the slider.

Also some devices need to press right key several times to change the text
shown on the config page. This CL also changes that to use testing.Poll()
to ensure the size is changing.

BUG=chromium:1105189
TEST=tast run -buildbundle=crostint graphics.TraceReplay.top10_2d

Change-Id: I51de9203222f6d045d2e04e7aa6140f68f3ef1c0