WIP: Introduce a mechanism to generate test parameters statically.
This change introduces generate.Params which takes two parameters:
a path of a Go source code defining a Tast test, and a Go expression
string representing []testing.Param that is expected to be found in
the source code. The function extracts a testing.Test literal from
the code and verifies that it matches with the expected parameters.
Also, when TAST_GENERATE_PARAMS environment variable is set to 1,
the function instead rewrites the source code to update Params with
the expected one.

generate.Params can be used to generate test parameters statically.

BUG=chromium:1065846
TEST=TBD

Change-Id: I874eb2c204d496defb84f03a994a0427a8d37e57