security.StatefulFiles: Mark as critical.
We've had 12 consecutive build with only two single failures, which is
fixed by:
https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2304253

And the big block of failures between 13331 and 13342 is fixed by:
https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2293248

https://tastboard.corp.google.com/test/security.StatefulFiles

Mark as critical now that we've had a few successful builds. This test
prevents security issues by ensuring that modifiable, persisted files
in stateful have the right permissions, but because of that it's also
easily broken if not run in presubmit.

BUG=chromium:1104857
TEST=Still passes on betty-pi-arc.

Change-Id: I6f234594723e6f25340206a86428449df85ec38f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2296275
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>