tast-tests: disable arc.IPv6Connectivity until lab supports IPv6
This test is failing 100% of the time because the lab does not currently
support IPv6.  This test must be run manually on a different network
which does support IPv6.

The whitespace changes were done by 'gofmt' in response to this error:

  [FAILED] chromiumos/platform/tast-tests: _check_gofmt
      Files not formatted with gofmt:
      src/chromiumos/tast/local/bundles/cros/arc/ipv6_connectivity.go

BUG=b:161239666
TEST=none

Change-Id: Idbf92564f1694f7711bd5461ca19f21da58f6f14
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2297234
Tested-by: Ross Zwisler <zwisler@google.com>
Auto-Submit: Ross Zwisler <zwisler@google.com>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Ross Zwisler <zwisler@google.com>