Add NT12 Tast-test for WM ARC++ R
Implement NV12:NT12: non-resizable/tablet: hide Shelf.
In NT12 two sets of activities are tested. A portrait + unspecified
and landscape + unspecified. Each activity starts in tablet mode and
depends on its orientation, the display should rotate to match with
the desired orientation of the activity. Then the shelf will be
hidden and shown and at each state the window bounds are stored to be
asserted at the end.
Unspecified activities are running on the same orientation as the
previous activity, so in case of portrait + unspecified, the
unspecified desired orientation is portrait.
For the purposes of this test, a new method is added to
wm_tablet_helper.go named TabletShelfHideShowHelper(). This method
will be used in RT12 due to similar logic to NT12.
BUG=b:140203139
TEST=tast run $DUT arc.WMNonresizableTablet

Change-Id: I7eecfe905977a854822a67a06f416cc562f7ce52
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2298096
Tested-by: Armen Keshishian <armenk@google.com>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Auto-Submit: Armen Keshishian <armenk@google.com>
Commit-Queue: Armen Keshishian <armenk@google.com>