tast-tests: hwsec: Add hwsec.RecreateUserVaultTPM1 test
hwsec.RecreateUserVaultTPM1 test is similar to hwsec.RecreateUserVault
local test, except that this test is specifically for TPMv1.2, while the
original local test is specifically for TPMv2.0. 2 versions of the same
test is required because the way we treat reboot in TPMv1.2 and TPMv2.0.

BUG=b:160260730
TEST=tast run $dut hwsec.RecreateUserVaultTPM1 on relm

Change-Id: I3d3cb19eb52854bf078a6e968b2de8dfbac240c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2299786
Tested-by: John L Chen <zuan@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Commit-Queue: John L Chen <zuan@chromium.org>