tast-tests: Promote arc.PhysicalKeyboard again
The test was disabled in https://crrev.com/c/2225199 because of
flakiness. Since then, stability improvements were made, and according
to http://tastboard/test/arc.PhysicalKeyboard it has been stable for
the last >20 builds. (Only infra related failures and precondition
failures were seen. arc.Booted sometimes fails because arc.PerAppDensity
sometimes crashes ARC and Tast is not properly recovering from it, but
promoting this to non-informational changes the execution order so it
should not be an immediate problem.)

BUG=b:157749001
TEST=CQ passes

Change-Id: I6985306e86b74ad51b8c297ecd20bdae8851db7c